working on crud operations on the rest of the models

+++ resources/views/livewire/events/delete.blade.php
@@ -14,12 +14,18 @@ new class extends Component {
     public function deleteEvent()
     {
         $this->event->delete();
-        
-        $this->dispatch('eventDeleted');
 
         Flux::toast(heading: 'Event deleted', text: 'Your event was deleted successfully', variant: 'danger');
 
-        Flux::modals()->close();
+        $url = request()->header('Referer');
+
+        if ($url === url()->route('calendar', [Auth::user()->username])) {
+            $this->dispatch('eventDeleted');
+            Flux::modals()->close();
+        } else {
+            Flux::modals()->close();
+            $this->redirectRoute('calendar', [Auth::user()->username], navigate: true);
+        }
     }
 }; ?>
 
@@ -28,20 +34,20 @@ new class extends Component {
         <div class="space-y-6">
             <div>
                 <flux:heading size="lg">Delete event?</flux:heading>
-    
+
                 <flux:subheading>
                     <p>You're about to delete the event {{ $event->name }}.</p>
                     <p>Are you sure you want to proceed?</p>
                 </flux:subheading>
             </div>
-    
+
             <div class="flex gap-2">
                 <flux:spacer />
-    
+
                 <flux:modal.close>
                     <flux:button variant="ghost">Cancel</flux:button>
                 </flux:modal.close>
-    
+
                 <flux:button type="submit" variant="danger">Delete event</flux:button>
             </div>
         </div>
